test: split moutingProps.attrs tests
